==> 81.ABFProcessing/07.Standard_AA_IntervalTime.Rmd <==
---
title: "Signal statistic"
author: "Chao Tang"
date: 'Report created: `r Sys.Date()`'
output: 
  html_document: 
    code_folding: "hide"
    toc: true
    toc_depth: 4
    toc_float: 
      collapsed: false
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      warning = FALSE, 
                      message = FALSE, 
                      fig.align = 'center')
knitr::opts_knit$set(root.dir = "/mnt/raid61/Personal_data/tangchao/AANanopore")
```

```{r}
library(IRanges)
library(parallel)
library(S4Vectors)
library(data.table)
```

```{r}
files <- list.files("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/02.SelectedSignals", "signal.remove.outliers18_1.Rds", full.names = T)
lapply(files, function(fi) {
  aat <- gsub(".signal.remove.outliers18_1.Rds", "", basename(fi))
  print(aat)
  if(file.exists(paste0("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/04.IntervalTime/", aat, "_IntervalTime.txt"))) return(NULL)
  sig <- readRDS(fi)
  sig <- split(sig, sig$File)
  SigInterTime <- lapply(sig, function(sigi) {
    print(nrow(sigi))
    setkey(sigi, StartTime)
    ol <- as.data.table(findOverlaps(sigi[, IRanges(StartTime * 10000, EndTime * 10000)], sigi[, IRanges(StartTime * 10000, EndTime * 10000)], type = "within"))[queryHits != subjectHits]
    if(nrow(ol) > 0) {
      ol <- unique(apply(ol, 1, sort)[1, ])
      sigi <- sigi[-c(ol)]
      setkey(sigi, StartTime)
    }
    InterTime <- data.table(File = sigi[, unique(File)], 
                            StartTime = sigi[, EndTime][-c(nrow(sigi))], 
                            EndTime = sigi[, StartTime][-c(1)], 
                            SigBefore = sigi[, ID][-c(nrow(sigi))], 
                            SigAfter = sigi[, ID][-c(1)], 
                            SigBeforeBaseMean = sigi[, BaseMean][-c(nrow(sigi))], 
                            SigAfterBaseMean = sigi[, BaseMean][-c(1)])
    
    filei <- sigi[, unique(File)]
    abf <- readRDS(paste0("./analysis/81.ABFProcessing/ABF/ABF_", filei, ".Rds"))
    L0 <- fread(list.files("./analysis/81.ABFProcessing/SelectedL0", sigi[, unique(file_id)], full.names = T))
    ssd <- abf[Sm >= L0[, min(BaseMean)] & Sm <= L0[, max(BaseMean)], sd(pA)]
    abf[, L0 := Sm >= L0[, min(BaseMean) - ssd] & Sm <= L0[, max(BaseMean) + ssd]]
    
    ValidRegion <- IRanges(Rle(abf[, round(mV) == 50]))
    Valid <- data.table(satrt = abf[start(ValidRegion), Time], end = abf[end(ValidRegion), Time])
    InterTime <- InterTime[StartTime %inrange% Valid & EndTime %inrange% Valid]
    InterTime2 <- mclapply(seq_len(nrow(InterTime)), function(i) {
      abf[Time > InterTime[i, StartTime] & Time < InterTime[i, EndTime], .(Time1 = diff(range(Time)), Time2 = mean(L0) * diff(range(Time)))]
    }, mc.cores = 10)
    InterTime2 <- do.call(rbind, InterTime2)
    InterTime <- cbind(InterTime, InterTime2)
    return(InterTime)
  })
  SigInterTime <- do.call(rbind, SigInterTime)
  fwrite(SigInterTime, paste0("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/04.IntervalTime/", aat, "_IntervalTime.txt"), sep = "\t", quote = F, row.names = F)
})
```

His1

```{r}
sig <- fread("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/03.GroupedSignals/His1.signal.txt")
sig <- split(sig, sig$File)
SigInterTime <- lapply(sig, function(sigi) {
  print(nrow(sigi))
  setkey(sigi, StartTime)
  ol <- as.data.table(findOverlaps(sigi[, IRanges(StartTime * 10000, EndTime * 10000)], sigi[, IRanges(StartTime * 10000, EndTime * 10000)], type = "within"))[queryHits != subjectHits]
  if(nrow(ol) > 0) {
    ol <- unique(apply(ol, 1, sort)[1, ])
    sigi <- sigi[-c(ol)]
    setkey(sigi, StartTime)
  }
  InterTime <- data.table(File = sigi[, unique(File)], 
                          StartTime = sigi[, EndTime][-c(nrow(sigi))], 
                          EndTime = sigi[, StartTime][-c(1)], 
                          SigBefore = sigi[, ID][-c(nrow(sigi))], 
                          SigAfter = sigi[, ID][-c(1)], 
                          SigBeforeBaseMean = sigi[, BaseMean][-c(nrow(sigi))], 
                          SigAfterBaseMean = sigi[, BaseMean][-c(1)])
  
  filei <- sigi[, unique(File)]
  abf <- readRDS(paste0("./analysis/81.ABFProcessing/ABF/ABF_", filei, ".Rds"))
  L0 <- fread(list.files("./analysis/81.ABFProcessing/SelectedL0", sigi[, unique(as.character(file_id))], full.names = T))
  ssd <- abf[Sm >= L0[, min(BaseMean)] & Sm <= L0[, max(BaseMean)], sd(pA)]
  abf[, L0 := Sm >= L0[, min(BaseMean) - ssd] & Sm <= L0[, max(BaseMean) + ssd]]
  
  ValidRegion <- IRanges(Rle(abf[, round(mV) == 50]))
  Valid <- data.table(satrt = abf[start(ValidRegion), Time], end = abf[end(ValidRegion), Time])
  InterTime <- InterTime[StartTime %inrange% Valid & EndTime %inrange% Valid]
  InterTime2 <- mclapply(seq_len(nrow(InterTime)), function(i) {
    abf[Time > InterTime[i, StartTime] & Time < InterTime[i, EndTime], .(Time1 = diff(range(Time)), Time2 = mean(L0) * diff(range(Time)))]
  }, mc.cores = 10)
  InterTime2 <- do.call(rbind, InterTime2)
  InterTime <- cbind(InterTime, InterTime2)
  return(InterTime)
})
SigInterTime <- do.call(rbind, SigInterTime)
fwrite(SigInterTime, paste0("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/04.IntervalTime/His1_IntervalTime.txt"), sep = "\t", quote = F, row.names = F)
```

His2

```{r}
sig <- fread("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/03.GroupedSignals/His2.signal.txt")
sig <- split(sig, sig$File)
SigInterTime <- lapply(sig, function(sigi) {
  print(nrow(sigi))
  setkey(sigi, StartTime)
  ol <- as.data.table(findOverlaps(sigi[, IRanges(StartTime * 10000, EndTime * 10000)], sigi[, IRanges(StartTime * 10000, EndTime * 10000)], type = "within"))[queryHits != subjectHits]
  if(nrow(ol) > 0) {
    ol <- unique(apply(ol, 1, sort)[1, ])
    sigi <- sigi[-c(ol)]
    setkey(sigi, StartTime)
  }
  InterTime <- data.table(File = sigi[, unique(File)], 
                          StartTime = sigi[, EndTime][-c(nrow(sigi))], 
                          EndTime = sigi[, StartTime][-c(1)], 
                          SigBefore = sigi[, ID][-c(nrow(sigi))], 
                          SigAfter = sigi[, ID][-c(1)], 
                          SigBeforeBaseMean = sigi[, BaseMean][-c(nrow(sigi))], 
                          SigAfterBaseMean = sigi[, BaseMean][-c(1)])
  
  filei <- sigi[, unique(File)]
  abf <- readRDS(paste0("./analysis/81.ABFProcessing/ABF/ABF_", filei, ".Rds"))
  L0 <- fread(list.files("./analysis/81.ABFProcessing/SelectedL0", sigi[, unique(as.character(file_id))], full.names = T))
  ssd <- abf[Sm >= L0[, min(BaseMean)] & Sm <= L0[, max(BaseMean)], sd(pA)]
  abf[, L0 := Sm >= L0[, min(BaseMean) - ssd] & Sm <= L0[, max(BaseMean) + ssd]]
  
  ValidRegion <- IRanges(Rle(abf[, round(mV) == 50]))
  Valid <- data.table(satrt = abf[start(ValidRegion), Time], end = abf[end(ValidRegion), Time])
  InterTime <- InterTime[StartTime %inrange% Valid & EndTime %inrange% Valid]
  InterTime2 <- mclapply(seq_len(nrow(InterTime)), function(i) {
    abf[Time > InterTime[i, StartTime] & Time < InterTime[i, EndTime], .(Time1 = diff(range(Time)), Time2 = mean(L0) * diff(range(Time)))]
  }, mc.cores = 10)
  InterTime2 <- do.call(rbind, InterTime2)
  InterTime <- cbind(InterTime, InterTime2)
  return(InterTime)
})
SigInterTime <- do.call(rbind, SigInterTime)
fwrite(SigInterTime, paste0("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/04.IntervalTime/His2_IntervalTime.txt"), sep = "\t", quote = F, row.names = F)
```
